src/main/webapp/app/entities/carte-bancaire/update/carte-bancaire-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize } from 'rxjs/operators';

import { CarteBancaireFormService, CarteBancaireFormGroup } from './carte-bancaire-form.service';
import { ICarteBancaire } from '../carte-bancaire.model';
import { CarteBancaireService } from '../service/carte-bancaire.service';

@Component({
  selector: 'jhi-carte-bancaire-update',
  templateUrl: './carte-bancaire-update.component.html',
})
export class CarteBancaireUpdateComponent implements OnInit {
  isSaving = false;
  carteBancaire: ICarteBancaire | null = null;

  editForm: CarteBancaireFormGroup = this.carteBancaireFormService.createCarteBancaireFormGroup();

  constructor(
    protected carteBancaireService: CarteBancaireService,
    protected carteBancaireFormService: CarteBancaireFormService,
    protected activatedRoute: ActivatedRoute
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ carteBancaire }) => {
      this.carteBancaire = carteBancaire;
      if (carteBancaire) {
        this.updateForm(carteBancaire);
      }
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const carteBancaire = this.carteBancaireFormService.getCarteBancaire(this.editForm);
    if (carteBancaire.id !== null) {
      this.subscribeToSaveResponse(this.carteBancaireService.update(carteBancaire));
    } else {
      this.subscribeToSaveResponse(this.carteBancaireService.create(carteBancaire));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<ICarteBancaire>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(carteBancaire: ICarteBancaire): void {
    this.carteBancaire = carteBancaire;
    this.carteBancaireFormService.resetForm(this.editForm, carteBancaire);
  }
}
